Added mechanism for drawing "2D lists", for use with indicators.
Fix depth bar to appear back in top right when window resizes (broken
in 1.1.5).

Fix compass and clino outer edges to "light up" when dragging with mouse
pointer outside (like they do in 1.0).

Reworked indicator drawing code in preparation for using lists more.


git-svn-id: file:///home/survex-svn/survex/branches/survex-1_1@3011 4b37db11-9a0c-4f06-9ece-9ab7cdaee568

+++ src/gfxcore.h
@@ -202,7 +202,7 @@ private:
     }
 
     int GetClinoOffset() const;
-    void DrawTick(wxCoord cx, wxCoord cy, int angle_cw);
+    void DrawTick(wxCoord cx, wxCoord cy, Double angle_cw);
     wxString FormatLength(Double, bool scalebar = true);
 
     void SkinPassage(const vector<XSect> & centreline);
